+++ core/src/routes/(modules)/options/div/flexbox-grid-flex-shrink.svelte
@@ -1,0 +1,111 @@
+<script>
+
+    import "../../../../app.css";
+    import Select from "../../../uicomponents/select.svelte";
+
+    /**
+     * @type boolean
+     */
+    export let loaded;
+
+    /**
+     * Actual class
+     * @type string
+     */
+    export let classInput;
+
+    /**
+     * Actual part of class
+     * @type string
+     */
+    export let property;
+
+
+    /**
+     * @typedef {Object} SelectOptions
+     * @property {string} value
+     * @property {string} name
+     * @property {string} info
+    */
+
+
+    /**
+     * Value of element
+     * @type Array<SelectOptions>
+     */
+    let propertyOptions = [
+        {name: "Default",   value: "",          info: "Default"},
+        {name: "Shrink",      value: "shrink",      info: "flex-shrink: 1;\nUse shrink to allow a flex item to shrink if needed"},
+        {name: "shrink 0",    value: "shrink-0",    info: "flex-shrink: 0;\nUse shrink-0 to prevent a flex item from shrinking"},
+
+        // {name: "",     value: "",     info: ""},
+       
+    ];
+
+    /**
+     * List of all classes
+     * @type Array<string>
+     */
+    let propertyList = [];
+
+    propertyOptions.forEach(elm => {
+        propertyList.push(elm.value);
+    });
+
+    import { createEventDispatcher } from 'svelte';
+
+    const dispatch = createEventDispatcher();
+
+    function updateClass() {
+        dispatch('updateClass');
+    }
+
+    /**
+     * Used for sub elements update control.
+     * @type string
+    */
+    export let elementDataLoaded;
+
+    $: loaded, (()=>{
+        if(loaded == true) {
+            elementDataLoaded.split(" ").forEach( cls => {
+
+                var currentClass = cls.trim();
+
+                propertyList.forEach( elm => {
+                    if(elm === currentClass) property = elm;
+                });
+
+
+            });
+        }
+    })();
+
+    $: property, (()=>{
+        if(loaded == true) {
+
+            var newClass = "";
+
+            classInput.split(" ").forEach( cls => {
+
+                if(propertyList.includes(cls.trim())){
+                    ///Do nothing - If included then should be removed
+                }else{
+                    newClass += " " + cls.trim();
+                }
+            });
+
+            classInput = newClass.trim() + " " + property;
+            updateClass();
+        }
+    })();
+
+
+
+</script>
+
+
+<div class="w-full flex flex-row justify-between h-8 align-middle items-center">
+    <span title="Utilities for controlling how flex items shrink.">Flex Shrink</span>
+    <Select options={propertyOptions} bind:value={property}/>
+</div>