+++ core/src/routes/(modules)/options/div/flexbox-grid-flex.svelte
@@ -1,0 +1,113 @@
+<script>
+
+    import "../../../../app.css";
+    import Select from "../../../uicomponents/select.svelte";
+
+    /**
+     * @type boolean
+     */
+    export let loaded;
+
+    /**
+     * Actual class
+     * @type string
+     */
+    export let classInput;
+
+    /**
+     * Actual part of class
+     * @type string
+     */
+    export let property;
+
+
+    /**
+     * @typedef {Object} SelectOptions
+     * @property {string} value
+     * @property {string} name
+     * @property {string} info
+    */
+
+
+    /**
+     * Value of element
+     * @type Array<SelectOptions>
+     */
+    let propertyOptions = [
+        {name: "Default",       value: "",              info: "Default"},
+        {name: "1",        value: "flex-1",        info: "flex: 1 1 0%;\nUse flex-1 to allow a flex item to grow and shrink as needed, ignoring its initial size"},
+        {name: "Auto",     value: "flex-auto",     info: "flex: 1 1 auto;\nUse flex-auto to allow a flex item to grow and shrink, taking into account its initial size"},
+        {name: "Initial",  value: "flex-initial",  info: "flex: 0 1 auto;\nUse flex-initial to allow a flex item to shrink but not grow, taking into account its initial size"},
+        {name: "None",     value: "flex-none",     info: "flex: none;\nUse flex-none to prevent a flex item from growing or shrinking"},
+
+        // {name: "",     value: "",     info: ""},
+       
+    ];
+
+    /**
+     * List of all classes
+     * @type Array<string>
+     */
+    let propertyList = [];
+
+    propertyOptions.forEach(elm => {
+        propertyList.push(elm.value);
+    });
+
+    import { createEventDispatcher } from 'svelte';
+
+    const dispatch = createEventDispatcher();
+
+    function updateClass() {
+        dispatch('updateClass');
+    }
+
+    /**
+     * Used for sub elements update control.
+     * @type string
+    */
+    export let elementDataLoaded;
+
+    $: loaded, (()=>{
+        if(loaded == true) {
+            elementDataLoaded.split(" ").forEach( cls => {
+
+                var currentClass = cls.trim();
+
+                propertyList.forEach( elm => {
+                    if(elm === currentClass) property = elm;
+                });
+
+
+            });
+        }
+    })();
+
+    $: property, (()=>{
+        if(loaded == true) {
+
+            var newClass = "";
+
+            classInput.split(" ").forEach( cls => {
+
+                if(propertyList.includes(cls.trim())){
+                    ///Do nothing - If included then should be removed
+                }else{
+                    newClass += " " + cls.trim();
+                }
+            });
+
+            classInput = newClass.trim() + " " + property;
+            updateClass();
+        }
+    })();
+
+
+
+</script>
+
+
+<div class="w-full flex flex-row flex-grow justify-between h-8 align-middle items-center">
+    <span title="Utilities for controlling how flex items wrap.">Flex</span>
+    <Select options={propertyOptions} bind:value={property}/>
+</div>